backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.120.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "uvicorn>=0.38.0",
]